requirements.txt
pandas==2.0.1
psycopg2_binary==2.9.3
python-dotenv==1.0.0
python-telegram-bot==13.13
schedule==1.1.0
